Extract helper methods to own file
