+++ jest.config.js
@@ -1,0 +1,4 @@
+module.exports = {
+  testPathIgnorePatterns: ["helpers.js"],
+  testRegex: "(/__tests__/.*|(\\.|/)(test|spec))\\.(jsx?|tsx?)$"
+}
